Chore: move find process name to a single part
